refactor: split into Cargo workspace with clean architecture layers
Restructure the single crate into a Cargo workspace with four layer
crates, following clean architecture:

- domain: AmbiguousTimeStrategy (value objects to be expanded later)
- usecase: TimezoneTranslator and TranslationError
- infrastructure: local timezone detection (TZ env, /etc/localtime,
  /etc/timezone)
- presentation: clap command definition, input validation, run() flow

The root tzt package is now a thin binary that delegates to
presentation::run(). Behavior is unchanged; code is moved, not
rewritten.

Additional mechanical fixes required by the move:

- CLI version now comes from CARGO_PKG_VERSION (workspace version)
- assert_cmd/predicates moved to dev-dependencies
- Makefile copied a non-existent binary name; fixed to target/release/tzt
- doc examples referencing private modules marked ignore (doctests run
  now that code lives in lib crates)
- two clippy warnings in infrastructure fixed (manual ok, needless
  return)
