+++ Cargo.toml
@@ -1,21 +1,44 @@
-[package]
-name = "tzt"
+[workspace]
+members = [
+    "crates/domain",
+    "crates/usecase",
+    "crates/infrastructure",
+    "crates/presentation",
+]
+
+[workspace.package]
 version = "0.3.1"
 edition = "2021"
 license = "MIT"
+repository = "https://github.com/shunsock/timezone_translator"
+
+[workspace.dependencies]
+chrono = "0.4"
+chrono-tz = "0.10"
+clap = "4.5"
+regex = "1.12.2"
+thiserror = "2.0.18"
+domain = { path = "crates/domain" }
+usecase = { path = "crates/usecase" }
+infrastructure = { path = "crates/infrastructure" }
+presentation = { path = "crates/presentation" }
+
+[package]
+name = "tzt"
+version.workspace = true
+edition.workspace = true
+license.workspace = true
 description = "simple command-line utility that converts a given time from one timezone to another."
 homepage = "https://github.com/shunsock/timezone_translator"
-repository = "https://github.com/shunsock/timezone_translator"
+repository.workspace = true
 documentation = "https://github.com/shunsock/timezone_translator/blob/main/README.md"
 readme = "README.md"
 keywords = ["timezone", "time", "converter", "translator", "cli"]
 categories = ["command-line-utilities"]
 
 [dependencies]
+presentation.workspace = true
+
+[dev-dependencies]
 assert_cmd = "2.1"
-chrono = "0.4"
-chrono-tz = "0.10"
-clap = "4.5"
 predicates = "3.1"
-regex = "1.12.2"
-thiserror = "2.0.18"
