+++ crates/domain/src/lib.rs
@@ -1,0 +1,3 @@
+pub mod ambiguous_time_strategy;
+
+pub use ambiguous_time_strategy::AmbiguousTimeStrategy;
