+++ crates/infrastructure/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "infrastructure"
+version.workspace = true
+edition.workspace = true
+license.workspace = true
+repository.workspace = true
+description = "Infrastructure layer of tzt: detects the local timezone from the system."
+publish = false
+
+[dependencies]
+
+[dev-dependencies]
+regex.workspace = true
