+++ crates/infrastructure/src/get_system_timezone_from_etc_timezone.rs
