+++ crates/infrastructure/src/lib.rs
@@ -1,4 +1,6 @@
 mod get_system_timezone_from_env_var_tz;
 mod get_system_timezone_from_etc_localtime;
 mod get_system_timezone_from_etc_timezone;
-pub(crate) mod local_timezone_string_provider;
+pub mod local_timezone_string_provider;
+
+pub use local_timezone_string_provider::provide_local_timezone_string;
