+++ crates/infrastructure/src/local_timezone_string_provider.rs
@@ -1,13 +1,13 @@
-use super::get_system_timezone_from_env_var_tz::EnvironmentVariableTzProvider;
-use super::get_system_timezone_from_etc_localtime::get_system_timezone_from_etc_localtime;
-use super::get_system_timezone_from_etc_timezone::get_system_timezone_from_etc_timezone;
+use crate::get_system_timezone_from_env_var_tz::EnvironmentVariableTzProvider;
+use crate::get_system_timezone_from_etc_localtime::get_system_timezone_from_etc_localtime;
+use crate::get_system_timezone_from_etc_timezone::get_system_timezone_from_etc_timezone;
 
 /// Returns the name of the local timezone as a `String`.
 ///
 /// # Examples
 ///
-/// ```
-/// let timezone = local_timezone_string();
+/// ```ignore
+/// let timezone = provide_local_timezone_string();
 /// println!("Local timezone: {}", timezone);
 /// ```
 ///
@@ -15,7 +15,7 @@ use super::get_system_timezone_from_etc_timezone::get_system_timezone_from_etc_t
 ///
 /// if environment variable TZ and following file links are not found,
 /// this function return panic
-pub(crate) fn provide_local_timezone_string() -> String {
+pub fn provide_local_timezone_string() -> String {
     // read environment variable TZ
     let env_var_tz: Option<String> = EnvironmentVariableTzProvider::new(None).get_env_var_tz();
     if let Some(env_var_tz) = env_var_tz {
