+++ crates/presentation/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "presentation"
+version.workspace = true
+edition.workspace = true
+license.workspace = true
+repository.workspace = true
+description = "Presentation layer of tzt: CLI definition, input validation, and output."
+publish = false
+
+[dependencies]
+domain.workspace = true
+usecase.workspace = true
+infrastructure.workspace = true
+chrono.workspace = true
+chrono-tz.workspace = true
+clap.workspace = true
+regex.workspace = true
+thiserror.workspace = true
