+++ crates/presentation/src/command/command_definition.rs
@@ -1,8 +1,8 @@
 use super::arguments::{
     ambiguous_time_strategy::ambiguous_time_strategy, from::from, time::time, to::to,
 };
-use crate::infrastructure::current_local_timezone_provider::local_timezone_string_provider::provide_local_timezone_string;
 use clap::Command;
+use infrastructure::provide_local_timezone_string;
 
 /// # About:
 /// Provides the command definition for the `tzt` command.
@@ -11,7 +11,7 @@ use clap::Command;
 /// `Command` struct containing the command definition.
 ///
 /// # Example:
-/// ```
+/// ```ignore
 /// use clap::ArgMatches;
 /// use command::command_definition::command_provider;
 /// let user_input: ArgMatches = command_provider().get_matches();
@@ -21,7 +21,7 @@ pub(crate) fn command_provider() -> Command {
     let now_str: &'static str = Box::leak(now.into_boxed_str());
 
     Command::new("tzt - Timezone Translator")
-        .version("0.3.1")
+        .version(env!("CARGO_PKG_VERSION"))
         .author("shunsock")
         .about("translate time from one timezone to another")
         .arg(time())
