+++ crates/presentation/src/command/validated_options/validated_user_inputs.rs
@@ -1,6 +1,6 @@
-use crate::command::validated_options::ambiguous_time_strategy::AmbiguousTimeStrategy;
 use chrono::NaiveDateTime;
 use chrono_tz::Tz;
+use domain::AmbiguousTimeStrategy;
 
 pub(crate) struct ValidatedCommandOptions {
     time: NaiveDateTime,
