+++ crates/presentation/src/validator/ambiguous_time_strategy_validator.rs
@@ -1,5 +1,5 @@
-use crate::command::validated_options::ambiguous_time_strategy::AmbiguousTimeStrategy;
 use crate::validator::validation_error::ValidationError;
+use domain::AmbiguousTimeStrategy;
 
 pub(super) fn validate_string_for_ambiguous_time_strategy(
     ambiguous_time_strategy_str: &str,
