+++ crates/usecase/src/translation_error.rs
@@ -2,7 +2,7 @@ use chrono::NaiveDateTime;
 use chrono_tz::Tz;
 
 #[derive(thiserror::Error, Debug)]
-pub(crate) enum TranslationError {
+pub enum TranslationError {
     #[error("Translation Error: Output time and timezone does not exist. Please check DST rules.")]
     TranslationError {
         time: NaiveDateTime,
