--- src/command/validated_options.rs
@@ -1,2 +1,0 @@
-pub(crate) mod ambiguous_time_strategy;
-pub(crate) mod validated_user_inputs;
