--- src/command/validated_options/ambiguous_time_strategy.rs
@@ -1,5 +1,0 @@
-#[derive(Debug, Clone, Copy, PartialEq)]
-pub(crate) enum AmbiguousTimeStrategy {
-    Earliest,
-    Latest,
-}
