--- src/infrastructure.rs
@@ -1,1 +1,0 @@
-pub(crate) mod current_local_timezone_provider;
